_scripts/requirements.txt
requests
youtube-transcript-api>=0.6.2
google-generativeai>=0.8.3
